+++ docs/getting-started/changelog.md
@@ -12,6 +12,8 @@ _During the beta period, these restrictions may be relaxed in the event of a mis
   - Removed ~400 baked-in Animista animations because they were causing ~200KB of bloat (they can still be used with custom keyframes)
   - Reworked animations into a separate module ([`@shoelace-style/animations`](https://github.com/shoelace-style/animations)) so it's more maintainable and animations are sync with the latest version of animate.css
   - Animation and easing names are now camelcase (e.g. `easeInOut` instead of `ease-in-out`)
+- Added initial E2E tests [#169](https://github.com/shoelace-style/shoelace/pull/169)
+- Fixed a bug where `sl-hide` would be emitted twice when closing an alert with `hide()`
 - Fixed a bug in `sl-color-picker` where the toggle button was smaller than the preview button in Safari
 
 ## 2.0.0-beta.25